+++ LitebaseSDK/src/java/litebase/SQLUpdateStatement.java
@@ -48,7 +48,7 @@ class SQLUpdateStatement extends SQLStatement
    /**
     * The array with the indexes of the parameters.
     */
-   byte[] paramIndexes;
+   short[] paramIndexes; // juliana@253_14: corrected a possible AIOBE if the number of parameters of a prepared statement were greater than 128.
 
    /**
     * An array that indicates if a parameters is defined or not.
@@ -84,7 +84,8 @@ class SQLUpdateStatement extends SQLStatement
       Vm.arrayCopy(parser.fieldNames, 0, fields, 0, nValues);
 
       // Allocates space for the list of the parameters. Worst case: all fields are parameters.
-      paramIndexes = new byte[nValues];
+   // juliana@253_14: corrected a possible AIOBE if the number of parameters of a prepared statement were greater than 128.
+      paramIndexes = new short[nValues];
       paramDefined = new boolean[nValues];
       
       while (--nValues >= 0)
@@ -359,7 +360,10 @@ class SQLUpdateStatement extends SQLStatement
       SQLValue value;
       SQLValue[] recordAux = record;
       SQLBooleanClause clause = whereClause;
-      byte[] paramIndexesAux = paramIndexes;
+      
+      // juliana@253_14: corrected a possible AIOBE if the number of parameters of a prepared statement were greater than 128.
+      short[] paramIndexesAux = paramIndexes; 
+      
       byte[] storeNullsAux = storeNulls;
       
       Convert.fill(paramDefined, 0, paramDefined.length, false); // Cleans the parameter values of the update clause.
